+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#5](https://github.com/LaxarJS/webpack-jasmine-html-runner-plugin/issues/5): fixed spec-runner naming, docs
+
 
 ## v0.4.0